backend/requirements.txt
ollama
chromadb
pdfplumber
langchain
langchain-core
langchain-ollama
langchain-community
langchain_text_splitters
unstructured
unstructured[all-docs]
fastembed
sentence-transformers
elevenlabs
fastapi
python-multipart
boto3
redis
python-dotenv
